Update vite build
This will be required for future commits to build properly. It just adds
content.js and service-worker.js to the build process and then deposits
them in the root of dist/. This is required for the extension, but
wouldn't be for a normal webpage which is why I'm having to do this by
hand.