Hide evidence tuner when analyse button is clicked
If the analyse evidence button is clicked while the evidence tuner is
open the evidence tuner doesn't update to the new value. This hides it
until a new cell is clicked on, and it will be up to date again. In the
future, reclicking analyse evidence should not change manually edited
evidence.