Move doUrlsMatch method to utils.js
Create a new utilities js file for common code across the project.

I realised that we will need the doUrlsMatch method again in App.vue to
implement the next feature, so rather than duplicate the code, I've
created this helper utilities file for now and future use to make sure
that we're not needlessly duplicating code.