Fix dragging hypotheses onto table
Saves the original url that claims were extracted from to decide if
dragged text should be added to hypotheses or evidence.

Before, all text dragged onto the ach table was getting adding to the
list of evidence. However, this doesn't make sense to do when dragging
text from the article that claims were originally extracted from.

Chris agreed that this behaviour is ideal. (2024-06-11)