Add vite-svg-loader to project
We will use this in future for importing icons.

I've spent quite a few hours now looking into a solution that lets us
use icons in a semantically correct way, while allowing easy resizing,
correct font coloring and loading quickly i.e. not importing a 1MB file
to show an icon on the fly.

Adding a new import (unfortunately) seems to be the only reasonable way
I can find to do it for now without really overcomplicating things. I
guess I'll have to live with it.