Fix tooltip on button
This allows the tooltip to appear when the cursor is hovering over the
button.

This took a *lot* of debugging. The reason the tooltip intantiator
wasn't happy with accepting tooltipElement.value is because
tooltipElement.value is not an html element in this case; it's a vue
component.

This means that even though we thought we were passing bootstrap.Tooltip
the html element, we actually weren't. Instead we need a way to get the
root element from the component. Fortunately that's what $el does for
us.

With this in place we're now actually passing the html element to
bootstrap.Tooltip and it's instantiating properly!