Fix bug in save and load
When loading from a file it was looking like any autocompleted values
were being wiped, and on top of that, the autocomplete button had
broken.

This is caused by the JSON serialiser converting the undefined values in
manualMatrix to nulls. Once they are considered nulls then they wrongly
take priority over the results in analysedMatrix every time.

This changes the creation of achMatrix to treat any null values and
undefined values as the same. Ideally in future we should switch to
storing nulls inside manualMatrix by default though so they're always
the same.